refactor: move `to_int` to abstract round abci
We need `to_int` to be re-usable.
